Achievement component is splitted into two, Rectangle and Square acheivement
