+++ app/frontend/src/Components/Achievement/Achievement/Achievement.module.scss
@@ -1,33 +1,32 @@
-@import "../../colors";
+@import "../../../colors";
 .achievement {
     display: flex;
-    flex-direction: column ;
     align-items: center;
-    justify-content: center;
     margin: 10px;
     padding: 10px;
     border: 1px solid #ccc;
     border-radius: 8px;
-    transition: border-color 0.3s;  
+    transition: border-color 0.3s;
+    width: 20em;
+    height: 6em;
     background-color: $violet-light-30;
-    width: 10em;
-    height: 10em;
   }
-  
+
   .achievement:hover {
     border-color: #007bff; /* Highlight border on hover */
+    width: 20em;
+    height: 6em;
   }
-  
+
   .icon {
     margin-right: 10px;
     border-radius: 10%;
-    height: 50%;
-    width: 50%;
+    width: 25%;
+    height: 100%;
   }
-  
+
   .achievement_content {
     flex: 1;
-
   }
   
   .hovered .achievementTitle {
@@ -36,7 +35,6 @@
   .achievementTitle {
     visibility: visible;
     font-size: 18px;
-    text-align: center  ;
   }
 
   .hovered .achievementGame {
@@ -52,7 +50,6 @@
     font-size: 14px;
     color: #555;
   }
-
   .hovered .floatingDescription {
     visibility: visible;
     color: #fff;
@@ -63,7 +60,6 @@
     z-index: 1;
   }
   .floatingDescription {
-
     visibility: hidden;
     color: #fff;
     text-align: center  ;
