+++ app/frontend/src/router.tsx
@@ -20,7 +20,7 @@ import GiveAdminPermission from "./Pages/Admin/User/GiveAdminPermission/GiveAdmi
 import Groups from "./Pages/Groups/Groups";
 import Group from "./Pages/Group/Group";
 import CreateGroup from "./Pages/CreateGroup/CreateGroup";
-import Achievement from "./Components/Achievement/Achievement";
+import Achievement from "./Components/Achievement/Achievement/Achievement";
 import Achievements from "./Pages/Achievement/Achievements";
 
 axios.defaults.headers.common["Content-Type"] = "application/json";
